Update current presets and harden GCSIM integration

+++ src/features/artifacts/scoringViewModel.ts
@@ -25,8 +25,8 @@ export interface ScoreBatchView {
   readonly buildStatus: Uint8Array;
   readonly buildIssueFlags: Uint32Array;
   readonly buildSetPlan: Uint8Array;
-  readonly match: Float64Array;
-  readonly expectedFinalMatch: Float64Array;
+  readonly match: Float32Array;
+  readonly expectedFinalMatch: Float32Array;
   readonly isPreferredMain: Uint8Array;
   readonly setCompatibility: Uint8Array;
   readonly pairIssueFlags: Uint32Array;
